youtube-trimmer/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Video processing
yt-dlp==2023.11.16
python-ffmpeg==2.0.12

# AI providers
openai==1.54.3
anthropic==0.7.7

# Configuration and environment
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.27.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# Payments
stripe==7.8.0

# Caching and rate limiting
redis==5.0.1
aioredis==2.0.1

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Additional middleware dependencies
slowapi==0.1.9
starlette==0.27.0